♻️ refactor :scss 컬러 변수 수정, 클래스네임 container -> 컴포넌트명-container로 수정
