+++ src/components/Common/Buttons/FollowButton/FollowButtonType.ts
@@ -1,3 +1,3 @@
 export default interface FollowButtonType {
   onClick: React.MouseEventHandler<HTMLButtonElement>;
-};
+}
