+++ src/components/Common/Buttons/NoticeButton/NoticeButtonType.ts
@@ -1,3 +1,3 @@
 export default interface NoticeButtonType {
   onClick: React.MouseEventHandler<HTMLButtonElement>;
-};
+}
