+++ src/components/Common/Buttons/SortDropdown/SortDropdownType.ts
@@ -2,4 +2,4 @@ export default interface SortDropdownType {
   onSortFollow: () => void;
   onSortAll: () => void;
   onSortMyGen: () => void;
-};
+}
