+++ src/components/Common/Layout/NavigationBar/NavigationBar.module.scss
@@ -1,9 +1,9 @@
-.container {
+.navigaionBar-container {
   width: 100%;
   max-width: 740px;
   height: 60px;
   padding: 12px 20px;
-  background: $gary-02;
+  background: $gary-01;
   display: flex;
   align-items: center;
   justify-content: space-between;
