+++ src/components/Common/Layout/SideBar/SideBar.module.scss
@@ -1,8 +1,8 @@
-.container {
+.sideBar-container {
   width: 100%;
   height: 72px;
   padding: 12px 24px;
-  background: $gary-01;
+  background: $gary-02;
   display: flex;
   position: fixed;
   bottom: 0;
@@ -23,7 +23,6 @@
     width: 76px;
     height: 268px;
     border-radius: 70px;
-    background: var(--gray2, #dbdbdb);
     backdrop-filter: blur(15px);
     padding: 30px 0;
     top: 0;
