+++ src/components/Common/PopOverBox/PopOver.module.scss
@@ -1,4 +1,4 @@
-.container {
+.popOver-container {
   width: 230px;
   margin: 0;
   padding: 20px 0;
